# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/reader": {
            "post": {
                "description": "Создаёт нового читателя с предоставленными данными.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readers"
                ],
                "summary": "Create a new reader",
                "parameters": [
                    {
                        "description": "Параметры для создания читателя. Пример: {\\",
                        "name": "reader",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_application_http_handlers_readers.CreateReaderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный читатель с уникальным ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reader/{id}": {
            "get": {
                "description": "Возвращает данные читателя по его уникальному идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readers"
                ],
                "summary": "Get a reader by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID читателя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные читателя",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Читатель не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные существующего читателя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readers"
                ],
                "summary": "Update a reader",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID читателя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные читателя. Пример: {\\",
                        "name": "reader",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_application_http_handlers_readers.UpdateReaderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённые данные читателя",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет читателя по его уникальному идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readers"
                ],
                "summary": "Delete a reader",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID читателя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Читатель успешно удалён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/readers": {
            "get": {
                "description": "Возвращает список всех читателей.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readers"
                ],
                "summary": "List all readers",
                "responses": {
                    "200": {
                        "description": "Список читателей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_application_http_handlers_readers.CreateReaderRequest": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean",
                    "example": false
                },
                "email": {
                    "type": "string",
                    "example": "ivan@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "phone": {
                    "type": "string",
                    "example": "+79111234567"
                }
            }
        },
        "internal_application_http_handlers_readers.UpdateReaderRequest": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean",
                    "example": false
                },
                "email": {
                    "type": "string",
                    "example": "ivan@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "password": {
                    "type": "string",
                    "example": "newpassword"
                },
                "phone": {
                    "type": "string",
                    "example": "+79111234567"
                }
            }
        },
...
    }
}